Only close zip file if set
When an invalid file is passed in and the Zip library throws an error,
@zip is not set. Then the ensure runs and tries to close the zip var,
which is nil. This raises a new error that covers up the actual error,
making the issue harder to investigate. By only closing the zip file if
it was actually opened, the original error is raised for invalid docx
files.